Moved tests to match production code
