+++ feature_album/src/test/java/com/igorwojda/showcase/feature/album/presentation/albumlist/AlbumListViewModelTest.kt
@@ -1,10 +1,8 @@
-package com.igorwojda.showcase.feature.album.presentation
+package com.igorwojda.showcase.feature.album.presentation.albumlist
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.igorwojda.showcase.feature.album.domain.model.AlbumDomainModel
 import com.igorwojda.showcase.feature.album.domain.usecase.GetAlbumListUseCase
-import com.igorwojda.showcase.feature.album.presentation.albumlist.AlbumListFragmentDirections
-import com.igorwojda.showcase.feature.album.presentation.albumlist.AlbumListViewModel
 import com.igorwojda.showcase.feature.album.presentation.albumlist.AlbumListViewModel.ViewState
 import com.igorwojda.showcase.library.base.presentation.navigation.NavigationManager
 import com.igorwojda.showcase.library.testutils.CoroutineRule
